Major changes
Talk about timestamps in requests more generally.  Date use is rare
(even if we are going to recommend it in OHTTP), but it is common to
have specific timestamp fields in things like signatures.

Add text about measuring server-client RTT using retries (a new tracking
vector).  And the way in which the client-server RTT might leak if the
client applies corrections based on the RTT.  This stops short of
recommending that clients not apply this correction, but we might
discuss whether that is a good recommendation.

Small cleanups throughout.